+++ mod/wikirate_csv_import/set/type/relationship_answer_import_file.rb
@@ -22,10 +22,6 @@ def import_item_class
   RelationshipAnswerImportItem
 end
 
-def item_label
-  "relationship answer"
-end
-
 # to get rid of the render errors that appear on relationship answer listings
 # on relationship metric pages on the first page load after an import
 event :clear_cache_after_relationship_import, after: :import_csv do
